Issues with updating ( see description)
So frappejs>client>views>form.js has the code for right layout, including save button.

I've tested the code the there of makeSendButton and it works.

Now again creating customised view, I do not understand how to make it work here. Are my objectNames the problem? Is it not exporting properly? or if I need to create seperate file for this, and if so how to link it to this file and make in work in accounting/client/index.js